Made correctionsto: remove .py~ files, py2/3 compatible string conversion, proper structure of project, added parsing funct. in parser.py
